rename: EmailAuthTimer 파일명 대소문자변경
